% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlugosc_luku.R
\name{dlugosc_luku}
\alias{dlugosc_luku}
\title{Dlugosc luku kola}
\usage{
dlugosc_luku(kolo, a)
}
\arguments{
\item{kolo}{zbudowany wczesniej obiekt o klasie "kolo" przy uzyciu
funkcji kolo.}

\item{a}{wartosc kata wycinka kola w stopniach}
}
\value{
wartosc numeryczna - dlugosc luku obiektu o klasie "kolo".
}
\description{
funkcja sluzacza do obliczenia dlugosci luku w kole.
Funkcja dziala tylko dla obiektow o klasie "kolo".
}
\examples{
dlugosc_luku(kolo(1, 9, 2, 10), 4)

fajne_kolo = kolo(pi, 300, pi+10, 310)
dlugosc_luku(fajne_kolo, 85)

dlugosc_luku(kolo(1, 9, 2, 10), 180)
dlugosc_luku(kolo(150, 150, 300, 300), 1)
}
