% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolo.R
\name{kolo}
\alias{kolo}
\title{Kolo}
\usage{
kolo(xmin, ymin, xmax, ymax)
}
\arguments{
\item{xmin}{minimalna wartosc wspolrzednej x}

\item{ymin}{minimalna wartosc wspolrzednej y}

\item{xmax}{maksymalna wartosc wspolrzednej x}

\item{ymax}{maksymalna wartosc wspolrzednej y}
}
\value{
obiekt o klasie "kolo"
}
\description{
funkcja tworzaca obiekt jakim jest kolo na podstawie
odpowiednich parametrow - wspolrzednych.
}
\examples{
kolo(2,5,10,13)

kolo(10,90,20,100)
}
