% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwadrat.R
\name{kwadrat}
\alias{kwadrat}
\title{Kwadrat}
\usage{
kwadrat(xmin, ymin, xmax, ymax)
}
\arguments{
\item{xmin}{najmniejsza wartosc polozenia na osi x}

\item{ymin}{najmniejsza wartosc polozenia na osi y}

\item{xmax}{najwieksza wartosc polozenia na osi x}

\item{ymax}{najwieksza wartosc polozenia na osi y}
}
\value{
obiekt o klasie "kwadrat"
}
\description{
funkcja tworzaca obiekt jakim jest kwadrat na podstawie
odpowiednich parametrow - wspolrzednych.
}
\examples{
kwadrat(2, 1, 5, 4)

kwadrat(4, 4, 192, 192)
}
