% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obwod_kolo.R
\name{obwod_kolo}
\alias{obwod_kolo}
\title{Obwod kola}
\usage{
obwod_kolo(kolo)
}
\arguments{
\item{kolo}{zbudowany wczesniej obiekt o klasie "kolo" przy uzyciu
funkcji kolo.}
}
\value{
wartosc numeryczna - obwod obiektu o klasie "kolo".
}
\description{
funkcja sluzacza do obliczenia obwodu kola.
Funkcja dziala tylko dla obiektow o klasie "kolo".
}
\examples{
obwod_kolo(kolo(10, 90, 20, 100))

ob = kolo(pi, 300, pi+10, 310)
obwod_kolo(ob)
}
