% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obwod_kwadrat.R
\name{obwod_kwadrat}
\alias{obwod_kwadrat}
\title{Obwod kwadratu}
\usage{
obwod_kwadrat(kwadrat)
}
\arguments{
\item{kwadrat}{zbudowany wczesniej obiekt o klasie "kwadrat" przy uzyciu
funckji kwadrat.}
}
\value{
wartosc numeryczna - obwod obiektu o klasie "kwadrat".
}
\description{
funkcja sluzacza do obliczenia obowdu kwadratu.
Funkcja dziala tylko dla obiektow o klasie "kwadrat".
}
\examples{
obwod_kwadrat(kwadrat(7, 8, 9, 10))

z = kwadrat(1, 1, 2, 2)
obwod_kwadrat(z)

fajny_kwadrat = kwadrat(0, 3, log10(10), log10(10000))
obwod_kwadrat(fajny_kwadrat)
}
