% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obwod_trojkat.R
\name{obwod_trojkat}
\alias{obwod_trojkat}
\title{Obwod trojkata}
\usage{
obwod_trojkat(trojkat)
}
\arguments{
\item{trojkat}{zbudowany wczesniej obiekt o klasie "trojkat" przy uzyciu
funkcji kwadrat.}
}
\value{
wartosc numeryczna - obwod obiektu o klasie "trojkat".
}
\description{
funkcja sluzacza do obliczenia obwodu trojkata.
Funkcja dziala tylko dla obiektow o klasie "trojkat".
}
\examples{
obwod_trojkat(trojkat(pi, pi + 1, 13, 14, 20, 40))

ob = trojkat(1, 2, 3, 1, 2, 3)
obwod_trojkat(ob)

}
