% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pole_prostokat.R
\name{pole_powierzchni_prostokat}
\alias{pole_powierzchni_prostokat}
\title{Pole powierzchni prostokata}
\usage{
pole_powierzchni_prostokat(x)
}
\arguments{
\item{x}{zbudowany wczesniej obiekt o klasie "prostokat" przy uzyciu
funckji prostokat.}
}
\value{
wartosc numeryczna - pole powierzchni obiektu o klasie "prostokat".
}
\description{
funkcja sluzacza do obliczenia pola powierzchni prostokata.
Funkcja dziala tylko dla obiektow o klasie "prostokat".
}
\examples{
pole_powierzchni_prostokat(prostokat(2, 3, 6, 6))

z = prostokat(0, 0, 12, 13)
pole_powierzchni_prostokat(z)

m = prostokat(log10(10), pi, 13, 204)
pole_powierzchni_prostokat(m)
}
