% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pole_trojkat.R
\name{pole_powierzchni_trojkat}
\alias{pole_powierzchni_trojkat}
\title{Pole powierzchni trojkata}
\usage{
pole_powierzchni_trojkat(x)
}
\arguments{
\item{x}{zbudowany wczesniej obiekt o klasie "trojkat" przy uzyciu
funkcji trojkat.}
}
\value{
wartosc numeryczna - pole powierzchni obiektu o klasie "trojkat".
}
\description{
funkcja sluzacza do obliczenia pola powierzchni trojkata
rownobocznego i rownoramiennego.
Funkcja dziala tylko dla obiektow o klasie "trojkat".
}
\examples{
pole_powierzchni_trojkat(trojkat(3, 0, 1, 6, 7, 5))

xc = trojkat(3, 0, 1, 6, 7, 5)
pole_powierzchni_trojkat(xc)

}
