% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pole_wycinka.R
\name{pole_wycinka}
\alias{pole_wycinka}
\title{Pole wycinka kola}
\usage{
pole_wycinka(kolo, a)
}
\arguments{
\item{kolo}{zbudowany wczesniej obiekt o klasie "kolo" przy uzyciu
funkcji kolo.}

\item{a}{wartosc kata wycinka kola w stopniach}
}
\value{
wartosc numeryczna - pole powierzchni wycinka kola
obiektu o klasie "kolo".
}
\description{
funkcja sluzacza do obliczenia pola wycinka kola.
Funkcja dziala tylko dla obiektow o klasie "kolo".
}
\examples{
pole_wycinka(kolo(1, 9, 2, 10), 4)

fajne_kolo = kolo(pi, 300, pi+10, 310)
pole_wycinka(fajne_kolo, 180)

pole_wycinka(kolo(1, 9, 2, 10), 180)
pole_wycinka(kolo(150, 150, 300, 300), 1)
}
