% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prostokat.R
\name{prostokat}
\alias{prostokat}
\title{Prostokat}
\usage{
prostokat(xmin, ymin, xmax, ymax)
}
\arguments{
\item{xmin}{najmniejsza wartosc polozenia na osi x}

\item{ymin}{najmniejsza wartosc polozenia na osi y}

\item{xmax}{najwieksza wartosc polozenia na osi x}

\item{ymax}{najwieksza wartosc polozenia na osi y}
}
\value{
obiekt o klasie "prostokat"
}
\description{
funkcja tworzaca obiekt jakim jest prostokat na podstawie
odpowiednich parametrow - wspolrzednych.
}
\examples{
prostokat(2, 3, 6, 6)

z = prostokat(1, 4, 12, 18)

prostokat(log10(10), pi, 144, 192)
}
