% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przekatna_kwadrat.R
\name{przekatna_kwadrat}
\alias{przekatna_kwadrat}
\title{Przekatna kwadratu}
\usage{
przekatna_kwadrat(x)
}
\arguments{
\item{x}{zbudowany wczesniej obiekt o klasie "kwadrat" przy uzyciu
funckji kwadrat.}
}
\value{
wartosc numeryczna - dlugosc przekatnej obiektu o klasie "kwadrat".
}
\description{
funkcja sluzacza do obliczenia dlugosci przekatnej kwadratu.
Funkcja dziala tylko dla obiektow o klasie "kwadrat".
}
\examples{
przekatna_kwadrat(kwadrat(1, 1, 4, 4))

kwadracik = kwadrat(4, 5, 6, 7)
przekatna_kwadrat(kwadracik)

d = kwadrat(log10(10), 2, log10(100), 3)
przekatna_kwadrat(d)
}
