% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przekatna_prostokat.R
\name{przekatna_prostokat}
\alias{przekatna_prostokat}
\title{Przekatna prostokata}
\usage{
przekatna_prostokat(x)
}
\arguments{
\item{x}{zbudowany wczesniej obiekt o klasie "prostokat" przy uzyciu
funckji prostokat.}
}
\value{
wartosc numeryczna - dlugosc przekatnej obiektu o klasie "prostokat".
}
\description{
funkcja sluzacza do obliczenia dlugosci przekatnej prostokata.
Funkcja dziala tylko dla obiektow o klasie "prostokat".
}
\examples{
przekatna_prostokat(prostokat(100, 101, 102, 103))

z = prostokat(4, 5, 6, 6)
przekatna_prostokat(z)

m = prostokat(log10(100), 1, 80, 2)
przekatna_prostokat(m)

}
