% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trojkat.R
\name{trojkat}
\alias{trojkat}
\title{Trojkat}
\usage{
trojkat(x1, x2, x3, y1, y2, y3)
}
\arguments{
\item{x1}{wspolrzedna pierwszego punktu na osi x}

\item{x2}{wspolrzedna drugiego punktu na osi x}

\item{x3}{wspolrzedna trzeciego punktu na osi x}

\item{y1}{wspolrzedna pierwszego punktu na osi y}

\item{y2}{wspolrzedna drugiego punktu na osi y}

\item{y3}{wspolrzedna trzeciego punktu na osi y}
}
\value{
obiekt o klasie "trojkat"
}
\description{
funkcja tworzaca obiekt jakim jest trojkat na podstawie
odpowiednich parametrow - wspolrzednych.
}
\examples{
trojkat(log(10), 2, 3, 100, 50, 12)

super_trojkat = trojkat(1, 4, 12, 18, 1, 7)

trojkat(log10(10), pi, 12^2, sqrt(7), 5, 94)
}
